requirements.txt
Django
hashids
django-hashid-field
